% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen2.R
\name{append_test}
\alias{append_test}
\title{Restructure and append tests to a test aggregating list}
\usage{
append_test(tag, tests, test)
}
\arguments{
\item{tag}{The roxygen tag that we are formatting}

\item{tests}{A \code{list} of test \code{code} objects to be formatted into a
\code{\\testonly} block.}

\item{test}{A new test \code{code} object to append to the list. If
necessary, the code will be modified to accommodate the testing style.}
}
\description{
Restructure and append tests to a test aggregating list
}
\seealso{
Other roclet_process_helpers: 
\code{\link{append_test_rd}()},
\code{\link{escape_infotex}()},
\code{\link{format_tests}()},
\code{\link{srcref_nlines}()}
}
\concept{roclet_process_helpers}
\keyword{internal}
