% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen2.R
\name{append_test_rd}
\alias{append_test_rd}
\title{Append a test to the examples Rd section}
\usage{
append_test_rd(ex, test)
}
\arguments{
\item{ex}{The existing character vector of example Rd section lines}

\item{test}{The additional test lines to add to the example}
}
\value{
The result of concatenating \code{test} into \code{ex}
}
\description{
Append a test to the examples Rd section
}
\note{
Because of how newlines are formatted when rendering Rd contents,
\verb{\testonly} blocks must starts on the last line of the code that they
test. Otherwise, an extra newline is printed when Rd is output as text.
}
\seealso{
Other roclet_process_helpers: 
\code{\link{append_test}()},
\code{\link{escape_infotex}()},
\code{\link{format_tests}()}
}
\concept{roclet_process_helpers}
\keyword{internal}
