% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{deparse_pretty}
\alias{deparse_pretty}
\title{Deparse pretty}
\usage{
deparse_pretty(expr)
}
\arguments{
\item{expr}{An expression to deparse}
}
\value{
A pretty-formatted string representation of \code{expr}.
}
\description{
Deparse to a single string with two spaces of indentation
}
\keyword{internal}
