% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen2.R
\name{escape_infotex}
\alias{escape_infotex}
\title{Escape escaped Rd \\testonly strings}
\usage{
escape_infotex(x)
}
\arguments{
\item{x}{A \code{character} value}
}
\description{
Escape escaped Rd \\testonly strings
}
\seealso{
Other roclet_process_helpers: 
\code{\link{append_test_rd}()},
\code{\link{append_test}()},
\code{\link{format_tests}()}
}
\concept{roclet_process_helpers}
\keyword{internal}
