% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testthat.R
\name{expect_no_example_error}
\alias{expect_no_example_error}
\title{testthat expectation asserting that code executes without error}
\usage{
expect_no_example_error(object, ...)
}
\arguments{
\item{object}{An expression to evaluate}

\item{...}{Additional arguments unused}
}
\description{
testthat expectation asserting that code executes without error
}
