% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testthat.R
\name{fallback_expect_no_error}
\alias{fallback_expect_no_error}
\title{Expect no Error}
\usage{
fallback_expect_no_error(object, ...)
}
\arguments{
\item{object}{An expression to evaluate}

\item{...}{Additional arguments unused}
}
\value{
The value produced by the expectation code
}
\description{
Expect no Error
}
\note{
This is a stop-gap implementation, and will only be used for legacy
versions of \code{testthat} before this was properly supported.

A \code{testthat} expectation that the provided code can be evaluated without
producing an error. This is the most basic expectation one should expect of
any example code. Further expectations are provided in subsequent \code{testthat}
code.
}
