% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{file_line_nchar}
\alias{file_line_nchar}
\title{Return the number of characters in a line of a file}
\usage{
file_line_nchar(file, line)
}
\arguments{
\item{file}{A file to use as reference}

\item{line}{A line number to retrieve the length of}
}
\description{
Return the number of characters in a line of a file
}
\keyword{internal}
