% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen2.R
\name{format_tests}
\alias{format_tests}
\title{Format test for an Rd \\testonly block}
\usage{
format_tests(tag, tests, ...)
}
\arguments{
\item{tag}{The roxygen tag that we are formatting}

\item{tests}{A \code{list} of test \code{code} objects to be formatted into a
\code{\\testonly} block.}
}
\description{
Format test for an Rd \\testonly block
}
\seealso{
Other roclet_process_helpers: 
\code{\link{append_test_rd}()},
\code{\link{append_test}()},
\code{\link{deparse_indent}()},
\code{\link{escape_infotex}()},
\code{\link{srcref_nlines}()}
}
\concept{roclet_process_helpers}
