% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testthat.R
\name{get_example_value}
\alias{get_example_value}
\title{Determine which symbol to use by default when testing examples}
\usage{
get_example_value()
}
\description{
Determine which symbol to use by default when testing examples
}
\keyword{internal}
