% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_r_cmd_check}
\alias{is_r_cmd_check}
\title{Test whether currently executing R CMD check}
\usage{
is_r_cmd_check()
}
\value{
A logical indicating whether \verb{R CMD check} is currently running
}
\description{
Test whether currently executing R CMD check
}
\keyword{internal}
