% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{package-file-helpers}
\alias{package-file-helpers}
\alias{find_package_root}
\alias{find_package_rds}
\alias{package_desc}
\title{Package source file helpers}
\usage{
find_package_root(path = ".", quiet = FALSE)

find_package_rds(package, path = getwd())

package_desc(path = getwd())
}
\arguments{
\item{path}{A file path within a package's source code or installation
directory. Only considered if \code{package} is missing.}

\item{quiet}{Whether to suppress output}

\item{package}{A package name}
}
\value{
NULL, invisibly

A list of package Rd objects, as returned by \code{\link[tools:Rdutils]{tools::Rd_db()}}
}
\description{
Discover specific package related file paths
}
\keyword{internal}
