% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.r
\name{parse.examples}
\alias{parse.examples}
\title{Parse examples.}
\usage{
parse.examples(sources)
}
\arguments{
\item{sources}{example sources defined by a call to \code{\link{testex_sources}}}
}
\description{
Parse examples.
}
