% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{priv}
\alias{priv}
\title{Extract select unexported objects from a package namespace}
\usage{
priv(package, names)
}
\arguments{
\item{package}{A package name}

\item{names}{A \code{character} vector of object names to select}
}
\description{
Extract select unexported objects from a package namespace
}
