% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen2.R
\name{roxy_test_helpers}
\alias{roxy_test_helpers}
\alias{roxy_test_raw_remainder}
\title{Separate the remaining text following a parsed roxygen test}
\usage{
roxy_test_raw_remainder(x)
}
\arguments{
\item{x}{A processed roxy testex tag, including a \verb{$test} field}
}
\value{
the string that follows the end of the test expression
}
\description{
Separate the remaining text following a parsed roxygen test
}
\keyword{internal}
