% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_srcref.R
\name{srclocs}
\alias{srclocs}
\title{Build a source location from a minimal numeric vector}
\usage{
srclocs(x, file)
}
\arguments{
\item{x}{A numeric vector of at least length 2}

\item{file}{A file to use to determine the length of the final line}
}
\value{
A numeric vector similar to a \code{\link[utils:sourceutils]{utils::getSrcLocation}} object
}
\description{
Build a length four source location from a length two source location. The
starting column on the first line is assumed to be 1, and the final column is
taken to be the length of the line if the source file exists, or 1 as a
fallback.
}
\keyword{internal}
