% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen2.R
\name{srcref_nlines}
\alias{srcref_nlines}
\title{Determine the number of source code lines of a given srcref}
\usage{
srcref_nlines(x)
}
\arguments{
\item{x}{A \code{srcref} object}
}
\description{
Determine the number of source code lines of a given srcref
}
\seealso{
Other roclet_process_helpers: 
\code{\link{append_test_rd}()},
\code{\link{append_test}()},
\code{\link{deparse_indent}()},
\code{\link{escape_infotex}()},
\code{\link{format_tests}()}
}
\concept{roclet_process_helpers}
