% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{string_line_count}
\alias{string_line_count}
\title{Return the number of lines in a multi-line string}
\usage{
string_line_count(x)
}
\arguments{
\item{x}{A character value}
}
\description{
Return the number of lines in a multi-line string
}
\keyword{internal}
