% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{string_newline_count}
\alias{string_newline_count}
\title{Get String Line Count}
\usage{
string_newline_count(x)
}
\arguments{
\item{x}{A character value}
}
\value{
The number of newline characters in a multiline string
}
\description{
Get String Line Count
}
\keyword{internal}
