% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testthat.R
\name{test_examples_as_testthat}
\alias{test_examples_as_testthat}
\title{Execute examples from Rd files as testthat tests}
\usage{
test_examples_as_testthat(
  package,
  path,
  ...,
  test_dir = tempfile("testex"),
  clean = TRUE,
  overwrite = TRUE,
  reporter = testthat::get_reporter()
)
}
\arguments{
\item{package}{A package name whose examples should be tested}

\item{path}{Optionally, a path to a source code directory to use. Will only
have an effect if parameter \code{package} is missing.}

\item{...}{Additional argument unused}

\item{test_dir}{An option directory where test files should be written.
Defaults to a temporary directory.}

\item{clean}{Whether the \code{test_dir} should be removed upon completion of test
execution. Defaults to \code{TRUE}.}

\item{overwrite}{Whether files should be overwritten if \code{test_dir} already
exists. Defaults to \code{TRUE}.}

\item{reporter}{A \pkg{testthat} reporter to use. Defaults to the active
reporter in the \pkg{testthat} environment or default reporter.}
}
\value{
The result of \code{\link[testthat:source_file]{testthat::source_file()}}, after iterating over
generated test files.
}
\description{
Reads examples from Rd files and constructs \pkg{testthat}-style tests.
\pkg{testthat} expectations are built such that
}
\details{
\enumerate{
\item Each example expression is expected to run without error
\item Any \code{testex} expectations are expected to pass
}
}
