% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testthat.R
\name{test_files}
\alias{test_files}
\title{Test a list of files}
\usage{
test_files(files, context, reporter, ...)
}
\arguments{
\item{files}{A collection of file paths to test}

\item{context}{An optional context message to display in \code{testthat} reporters}

\item{reporter}{A reporter to use when running each test file}

\item{...}{Additional arguments passed to \code{testhat::source_file}}
}
\value{
The result of \code{\link[testthat:source_file]{testthat::source_file()}}, after iterating over
generated test files.
}
\description{
Test a list of files
}
\keyword{internal}
