% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{testex-options}
\alias{testex-options}
\alias{memoise_testex_desc}
\alias{testex_options}
\title{Cached retrieval of testex options from package DESCRIPTION}
\usage{
memoise_testex_desc(path, fingerprint, ...)

testex_options(path = package_desc(), ...)
}
\arguments{
\item{path}{A path in which to search for a package \code{DESCRIPTION}}

\item{fingerprint}{An object used to indicate when the cached values have
been invalidated}
}
\value{
The test options environment, invisibly.

The test options environment as a list
}
\description{
As long as the \code{fingerprint} has not changed, the package \code{DESCRIPTION} will
be read only once to parse and retrieve configuration options. If the
\code{DESCRIPTION} file is modified or if run from a separate process, the
configured settings will be refreshed based on the most recent version of
the file.
}
\section{Functions}{
\itemize{
\item \code{testex_options()}: 

}}
\keyword{internal}
