% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{testex-private-imports}
\alias{testex-private-imports}
\alias{priv}
\alias{.roxygen2}
\title{Extract select unexported objects from a package namespace}
\usage{
priv(package, names)

.roxygen2()
}
\arguments{
\item{package}{A package name}

\item{names}{A \code{character} vector of object names to select}
}
\value{
A list of specified private namespace objects
}
\description{
Extract select unexported objects from a package namespace

Private roxygen2 functions
}
\keyword{internal}
