% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen2.R
\name{testex-roclets}
\alias{testex-roclets}
\title{testex replacement for roxygen2 rd roclet}
\value{
A new \code{roxygen2} \code{"rd"} roclet.
}
\description{
This roclet aims to be feature compatible with \pkg{roxygen2}'s \code{"rd"}
roclet. In addition it supports two new \code{roxygen} tags, \code{@expect}
and \code{@testthat}.
}
\section{tags}{

\code{testex} tags are all sub-tags meant to be used within an
\code{@examples} block. They should be considered as tags \emph{within} the
\code{@examples} block and used to construct blocks of testing code within
example code.

\describe{
\item{\code{@expect}: }{
In-line expectations to test the output of the previous command within an
example. If \code{.} is used within the expecation, it will be used to
refer to the output of the previous example command. Otherwise, the
result of the expression is expected to be identical to the previous
output.

\if{html}{\out{<div class="sourceCode">}}\preformatted{#' @examples
#' 1 + 2
#' @expect 3
#' @expect . == 3
#'
#' @examples
#' 3 + 4
#' @expect identical(., 7)
}\if{html}{\out{</div>}}

}
}

\describe{
\item{\code{@testthat}: }{
Similar to \code{@expect}, \code{@testthat} can be used to make in-line
assertions using \pkg{testthat} expectations. \pkg{testthat} expectations
follow a convention where the first argument is an object to compare
against an expected value or characteristic. Since the value will always
be the result of the previous example, this part of the code is
implicitly constructed for you.

If you want to use the example result elsewhere in your expectation, you
can refer to it with a \code{.}. When used in this way, \pkg{testex} will
not do any further implicit modification of your expectation.

\if{html}{\out{<div class="sourceCode">}}\preformatted{#' @examples
#' 1 + 2
#' @testthat expect_equal(3)
#' @testthat expect_gt(0)
#'
#' @examples
#' 3 + 4
#' @testthat expect_equal(., 7)
}\if{html}{\out{</div>}}

}
}
}

