% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testex.r
\name{testex_create}
\alias{testex_create}
\title{Create new example tests}
\usage{
testex_create(sources, exceptions = testex_exceptions(),
  parent.env = parent.frame(), verbose = TRUE)
}
\arguments{
\item{sources}{example sources defined by a call to \code{\link{testex_sources}}}

\item{exceptions}{possible exceptions of function calls or returned classes that shall not be compared, generated with \code{\link{testex_exceptions}}.}

\item{parent.env}{The parent environment in which examples are evaluated. By default \code{parent.frame()}.}

\item{verbose}{Shall extra information be shown while creating the tests?}
}
\description{
Create new example tests
}
