% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exceptions.r
\name{testex_exceptions}
\alias{testex_exceptions}
\title{Create exceptions, i.e. functions or classes that will not be compared}
\usage{
testex_exceptions(funs = NULL, classes = NULL, ignore.error.funs = NULL,
  global.ignore.funs = default.global.ignore.funs(),
  global.ignore.error.funs = NULL)
}
\description{
Create exceptions, i.e. functions or classes that will not be compared
}
