% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testex.r
\name{testex_run}
\alias{testex_run}
\title{Run example tests}
\usage{
testex_run(et, log.file = "example_test_log.Rmd",
  stat.file = "example_test_stats.csv", parent.env = parent.frame(),
  exceptions = et$exceptions)
}
\arguments{
\item{et}{An example test originally created with \code{\link{testex_create}}}

\item{log.file}{The file name of the log. Should be an Rmd file.}

\item{stat.file}{The name of the csv file that contains statistics about how many function calls of particular type failed.}

\item{parent.env}{The parent environment in which examples are evaluated.}
}
\description{
Reruns all examples that have been originally evaluated and stored in an example test.
}
\details{
Returns the number of calls that now return different values or throw an error and writes a detailed log as Rmd file.
}
