% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.r
\name{testex_sources}
\alias{testex_sources}
\title{Specify sources files or texts from which examples shall be extracted.}
\usage{
testex_sources(ex.in.fun.files = NULL, ex.files = NULL, rmd.files = NULL,
  ex.in.fun.text = NULL, ex.text = NULL)
}
\arguments{
\item{ex.in.fun.files}{Files that contain one or several examples within functions that have no argument. E.g. a function named \code{my.example()} whose body contains the examples.}

\item{ex.files}{Files whose code directly specifies an examples, i.e. examples are not within functions.}

\item{rmd.files}{Rmd files whose chunks specify examples. Chunks will be evaluated sequentially step by step. Chunks with \code{include=FALSE} or \code{eval=FALSE} will be ignored.}

\item{ex.in.fun.text}{Similar to \code{ex.in.fun.files} but code will be provided as string.}

\item{ex.text}{Similar to \code{ex.files} but code will be provided as string.}
}
\description{
There are different sources that differ how examples are written in the file
}
