% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testthat.R
\name{testthat_expect}
\alias{testthat_expect}
\alias{testthat_block}
\title{Helpers for using testex with testthat}
\usage{
testthat_expect(..., val, envir = parent.frame())

testthat_block(..., val, envir = parent.frame())
}
\description{
A flavor of \code{testex} that will inject \code{.Last.value} into the first
argument of each expression - suitable for using the \code{expect_*} family
of functions from \code{testthat}. Also handles temporarily attaching the
\code{testthat} package.

A flavor of \code{testex} that will inject \code{.Last.value} into the first
argument of each expression - suitable for using the \code{expect_*} family
of functions from \code{testthat}. Also handles temporarily attaching the
\code{testthat} package.
}
