% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.R
\name{use_rd_roclet}
\alias{use_rd_roclet}
\title{Replace default rd roclet with testex rd roclet}
\usage{
use_rd_roclet(path = getwd(), check = NA)
}
\arguments{
\item{path}{A package source code working directory}

\item{check}{A \code{logical} value indicating whether tests should be
executing during \code{R CMD check}.}
}
\value{
The result of \code{\link[=write.dcf]{write.dcf()}} upon modifying the package
\code{DESCRIPTION} file.
}
\description{
Replace default rd roclet with testex rd roclet
}
\note{
The testex roclet aims to be functionally identical to the default roxygen rd
roclet for any default roxygen tags. It replaces the default to intersperse
tests in the midst of existing \code{\\examples{...}} sections of Rd files.
}
