% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{uses_roxygen2}
\alias{uses_roxygen2}
\title{Checks for use of \code{roxygen2}}
\usage{
uses_roxygen2(path)
}
\arguments{
\item{path}{A file path to a package source code directory}
}
\value{
A logical value indicating whether a package takes \code{roxygen2} as
a dependency.
}
\description{
Checks for use of \code{roxygen2}
}
