% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{with_attached}
\alias{with_attached}
\title{Temporarily attach a namespace}
\usage{
with_attached(ns, expr)
}
\arguments{
\item{ns}{A namespace or namespace name to attach}

\item{expr}{An expression to evaluate while the namespace is attached}
}
\description{
Temporarily attach a namespace
}
