% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testthat.R
\name{wrap_expect_no_error}
\alias{wrap_expect_no_error}
\title{Wraps an example expression in a testthat expectation to not error}
\usage{
wrap_expect_no_error(expr, value)
}
\description{
Wraps an example expression in a testthat expectation to not error
}
\keyword{internal}
