% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_srcref.R
\name{as.srcref}
\alias{as.srcref}
\alias{as.srcref.character}
\title{Convert to srcref}
\usage{
as.srcref(x)

\method{as.srcref}{character}(x)
}
\arguments{
\item{x}{an object to coerce}
}
\value{
A \code{srcref}
}
\description{
Convert to srcref
}
\section{Methods (by class)}{
\itemize{
\item \code{as.srcref(character)}: Convert from a \code{srcref_key} to a sourceref object

}}
\keyword{internal}
