% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen2.R
\name{deparse_indent}
\alias{deparse_indent}
\title{Deparse an expression and indent for pretty-printing}
\usage{
deparse_indent(x, indent = 0L)
}
\arguments{
\item{x}{A \code{code} object}

\item{indent}{An \code{integer} number of spaces or a string to prefix each
line of the deparsed output.}
}
\description{
Deparse an expression and indent for pretty-printing
}
\seealso{
Other roclet_process_helpers: 
\code{\link{append_test_rd}()},
\code{\link{append_test}()},
\code{\link{escape_infotex}()},
\code{\link{format_tests}()},
\code{\link{srcref_nlines}()}
}
\concept{roclet_process_helpers}
