% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_srcref.R
\name{split_srcref}
\alias{split_srcref}
\title{Split a srcref into separate srcrefs at specific lines}
\usage{
split_srcref(sr, where)
}
\arguments{
\item{sr}{An original srcref object}

\item{where}{A numeric vector of line numbers where the srcref should be
split}
}
\value{
A list of \code{srcref}
}
\description{
Split a srcref into separate srcrefs at specific lines
}
\keyword{internal}
