% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_srcref.R
\name{srcref_key}
\alias{srcref_key}
\title{Convert a srcref to a character representation}
\usage{
srcref_key(x, nloc = 2, path = c("base", "root", "full"))
}
\arguments{
\item{x}{A srcref object}

\item{nloc}{The number of src locations to use. Defaults to 2, indicating
starting and ending line number.}

\item{path}{A form of filepath to use for the key. One of \code{"base"} for only
the basename of the source filepath, \code{"root"} for a path relative to a
package root directory if found, or \code{"full"} for the full filepath.}
}
\value{
A string hash of a \code{srcref}
}
\description{
Convert a srcref to a character representation
}
\keyword{internal}
