% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_srcref.R
\name{srcref_nlines}
\alias{srcref_nlines}
\title{Determine the number of source code lines of a given srcref}
\usage{
srcref_nlines(x)
}
\arguments{
\item{x}{A \code{srcref} object}
}
\value{
The number of lines in the original source of a \code{srcref}
}
\description{
Determine the number of source code lines of a given srcref
}
\keyword{internal}
