% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{string_line_count}
\alias{string_line_count}
\title{Get String Line Count}
\usage{
string_line_count(x)
}
\arguments{
\item{x}{A character value}
}
\value{
The number of lines in a mult-line string
}
\description{
Get String Line Count
}
\keyword{internal}
