% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testthat.R
\name{test_examples_as_testthat}
\alias{test_examples_as_testthat}
\title{Execute examples from Rd files as testthat tests}
\usage{
test_examples_as_testthat(
  package,
  path,
  ...,
  test_dir = tempfile("testex"),
  quiet = TRUE,
  clean = TRUE,
  overwrite = TRUE,
  reporter = testthat::get_reporter()
)
}
\arguments{
\item{package}{A package name whose examples should be tested}

\item{path}{Optionally, a path to a source code directory to use. Will only
have an effect if parameter \code{package} is missing.}

\item{...}{Additional argument unused}

\item{reporter}{A \pkg{testthat} reporter to use. Defaults to the active
reporter in the \pkg{testthat} environment or default reporter.}
}
\description{
Reads examples from Rd files and constructs \pkg{testthat}-style tests. Each
expression is expected to run without error and any in-line expectations
naturally result in individual expectations.
}
