% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testthat.R
\name{test_files}
\alias{test_files}
\title{Test a list of files}
\usage{
test_files(files, context, ...)
}
\arguments{
\item{files}{An iterable collection of file paths to test}

\item{context}{An optional context message to display in testthat reporters}

\item{...}{Additional arguments passed to \code{testhat::source_file}}
}
\description{
Test a list of files
}
\keyword{internal}
