% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_rd.R
\name{testex-rd-example-helpers}
\alias{testex-rd-example-helpers}
\alias{rd_extract_examples}
\alias{rd_example_as_string}
\alias{split_testonly_as_expr}
\alias{split_testonly}
\title{Rd Example Parsing Helpers}
\usage{
rd_extract_examples(rd)

rd_example_as_string(rd)

split_testonly_as_expr(rd)

split_testonly(rd)
}
\arguments{
\item{rd}{An Rd object}
}
\description{
Rd Example Parsing Helpers
}
\section{Functions}{
\itemize{
\item \code{rd_extract_examples()}: Extract examples tag from an Rd file

\item \code{rd_example_as_string()}: Convert an Rd example to string

\item \code{split_testonly_as_expr()}: Split sections of an example into evaluated example code blocks and code
blocks wrapped in testonly \code{Rd_tag}s, reassigning \code{srcref}s as the example
code is split.

\item \code{split_testonly()}: Split sections of an example into lists of \code{Rd_tag}s

}}
\keyword{internal}
