% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testex.R
\name{testex}
\alias{testex}
\title{A syntactic helper for writing quick and easy example tests}
\usage{
testex(
  ...,
  value = get_example_value(),
  obj = NULL,
  example = NULL,
  tests = NULL,
  envir = parent.frame()
)
}
\arguments{
\item{...}{Expressions to evaluated. \code{.} will be replaced with the
expression passed to \code{val}, and may be used as a shorthand for the
last example result.}

\item{value}{A value to test against. By default, this will use the example's
\code{.Last.value}.}

\item{obj}{An optional object name used to construct a more helpful error
message testthat failure message.}

\item{example}{An option \code{srcref_key} string used to indicate where the
relevant example code originated from.}

\item{tests}{An option \code{srcref_key} string used to indicate where the
relevant test code originated from.}

\item{envir}{An environment in which tests should be evaluated. By default
the parent environment where tests are evaluated.}
}
\value{
invisibly returns the \code{.Last.value} as it existed prior to evaluating
the test
}
\description{
A wrapper around \code{stopifnot} that allows you to use \code{.} to refer to
\code{.Last.value} and preserve the last non-test output from an example.
}
