% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vlapply}
\alias{vlapply}
\alias{vcapply}
\alias{vnapply}
\title{\code{vapply} shorthands}
\usage{
vlapply(..., FUN.VALUE = logical(1L))

vcapply(..., FUN.VALUE = character(1L))

vnapply(..., FUN.VALUE = numeric(1L))
}
\arguments{
\item{...}{Arguments passed to \code{\link{vapply}}}

\item{FUN.VALUE}{A preset signature for the flavor of \code{\link{vapply}}. This is
exposed for transparency, but modifying it would break the implicit
contract in the function name about the return type.}
}
\value{
The result of \code{\link{vapply}}
}
\description{
Simple wrappers around \code{vapply} for common data types
}
\keyword{internal}
